% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATED.R
\name{DATED}
\alias{DATED}
\title{DATED
Function to conduct hierarchical mutiple testing based on the leaf p-values}
\usage{
DATED(struct_map, L = 5, alpha = 0.05)
}
\arguments{
\item{struct_map}{a data frame with both leaf information and P-value information.}

\item{L}{maximum number of layers}

\item{alpha}{desired FDR}
}
\value{
a data frame with testing results.
}
\description{
DATED
Function to conduct hierarchical mutiple testing based on the leaf p-values
}
