% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_leaf.R
\docType{data}
\name{p_leaf}
\alias{p_leaf}
\title{p_leaf}
\format{
A data frame with 16281 rows and 5 variables. Each row links to a SNP that belongs to a leaf with testing p-value<1.
}
\usage{
p_leaf
}
\description{
p_leaf
}
\keyword{datasets}
